Create README.md
Website Mirroring Script with HTTrack
This Python script is designed to automate the process of mirroring websites using HTTrack, a free, open-source website copier that allows for offline browsing. The script enables users to specify the URL of the website they wish to mirror, the depth of the mirror (i.e., how many levels deep to follow links from the initial page), and the output directory where the mirrored website will be saved.